v2/devel/backend/pyproject.toml


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]

name = "amisr_system_monitor"
dynamic = ["version"]
description = "Real-time web display of AMISR system performance"
readme = "README.rst"
requires-python= ">= 3.9"
license = { file = "LICENSE" }
authors = [
    { name= "Todd Valentic", email="todd.valentic@gmail.com" }
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]

dependencies = [
    "pytz",
    "dateutils",
    "eventlet",
    "requests",
    "bcrypt",
    "flask",
    "psycopg2-binary",
    "sqlalchemy~=1.4",
    "Flask-SQLAlchemy",
    "flask-compress",
    "flask-script",
    "flask-migrate",
    "flask-bcrypt",
    "flask-jwt-extended",
    "flask-socketio",
    "flask-restful",
    "flask-mail",
    "flask-flatpages",
    "coverage",
    "PyJWT",
    "cryptography",
    "itsdangerous",
    "docutils"
]

[project.optional-dependencies]

test = [
    "pytest",
    "pyfakefs"
    ]

devel = [
    "build",
    "pylint",
    "black",
    "pip-tools"
    ]

[project.urls]

homepage="https://github.com/valentic/amisr_system_monitor"
repository="https://github.com/valentic/amisr_system_monitor"

[project.scripts]

server-cmd = "server.commands:cli"
server-launch = "server.commands:launch"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "server.metadata.version"}

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib"
]

